Module pallet_fennel_identity::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

DefaultCurrent

Pallet

The pallet implementing the on-chain logic.

_GeneratedPrefixForStorageFennelSignal

_GeneratedPrefixForStorageIdentityList

_GeneratedPrefixForStorageIdentityNumber

_GeneratedPrefixForStorageIdentityTraitList

_GeneratedPrefixForStorageRevokedIdentityNumber

_GeneratedPrefixForStorageSignalCount

Enums

Call

Contains one variant per dispatchable that can be called by an extrinsic.

Error

Custom dispatch errors of this pallet.

Event

The event emitted by this pallet.

Traits

Config

Configuration trait of this pallet.

Functions

__type_value_for_default_current

Type Definitions

FennelSignal

Tracks all signals sent by an identity.

IdentityList

Maps accounts to the array of identities it owns.

IdentityNumber

Tracks the number of identities currently active on the network.

IdentityTraitList

Maps identity ID numbers to their key/value attributes.

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

RevokedIdentityNumber

Tracks the number of identities currently active on the network.

SignalCount

Tracks the number of signals transmitted to the network.